23 Jul 03:15
i uh... wanted to go to bed early today.
i guess i didn't huh?

anyway i started the basics for elements and layout. still have to make the slider and checkboxes, as well as all the JS.

for now, it's mostly just an idea for how it'll look and function